Rename ledger_oasis_go package to internal and move it to its directory
