+++ internal/common_test.go
@@ -1,4 +1,4 @@
-package ledger_oasis_go
+package internal
 
 import (
 	"encoding/base64"
